third_party/WebKit/LayoutTests/fast/mediastream/RTCPeerConnection-createOffer-expected.txt

Tests RTCPeerConnection createOffer.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS requestSucceeded was called.
PASS sessionDescription.type is "offer"
PASS requestFailed was called.
PASS errorReason is "TEST_ERROR"
PASS pc.createOffer(dummy, dummy, {offerToReceiveVideo:1, offerToReceiveAudio:0, voiceActivityDetection:false, iceRestart:true}); did not throw exception.
PASS successfullyParsed is true

TEST COMPLETE
